Implemented fix for openSUSE Tumbleweed filename issues.
Resolved issues with the dynamic naming of openSUSE Tumbleweed non-vm images.

Previously, the Tumbleweed images used a frequently changing "SnapshotYYYYMMDD.iso" naming convention. With this commit, we default to the more consistent "Current.iso" scheme. I've added a step post-GPG verification: for standard Tumbleweed images, the SHA file is read to retrieve and set the actual "Snapshot" filename in "$iso_file", ensuring the correct filename is used for downloading and checksum verification.