Switch to dedicated GPG keyring instead of using user's.
This commit introduces a QoL change. Instead of using the user's GPG keyring and polluting it with various distribution keys, we are now using a dedicated GPG keyring for dl-distro.

The directories for the keyring are as follows:

- $XDG_DATA_HOME/dl-distro
- $HOME/.local/share/dl-distro

dl-distro should take care of everything, but as always, if an error occurs during fetching from the keyserver, manual intervention is required, or skipping verification. Updated 'Error Importing Signing Key' section of README.md.

A new option is also introduced with this commit:

-l, --listkeys

This will list the keys saved in the dedicated keyring.