Add netboot tarball for Ubuntu Mantic, JSON change
NOTE: This commit introduces a change to the `data.json` file that will
require users to update `dl-distro`.

JSON changes:
    .ubuntu.current -> .ubuntu.mantic

    .ubuntu.lts -> .ubuntu.jammy

These changes are to prepare for the upcoming commits,
introducing support for all available versions of Ubuntu currently.

Ubuntu flavors, such as, Kubuntu, will still be unsupported
for the time being.