dl-distro: ubuntu: fix trusty tahr gpg edge case
The Trusty Tahr images are signed with two keys.

The 2004 Automatic Signing key:
C5986B4F1257FFA86632CBA746181433FBB75451

The 2012 Automatic Signing key:
843938DF228D22F7B3742BC0D94AA3F0EFE21092

Trying to verify with just the 2012, as was the case previously,
resulted in an exit code of 2 from gpg. This resulted in an error
and file removal during verification.

This commit introduces a fix for this weird edge case.
If the user is downloading a Trusty Tahr image, dl-distro
will also import the 2004 Automatic Signing key.