Fedora: Implement logic for Astronomy GPG.
Since Astronomy is staying on Fedora 38,
this logic of fetching the old key must be introduced.

This is similar to Ubuntu Trusty,
also refactoring that a tad in this commit.

Note: Fedora Security is also on Fedora 38 currently.
Logic to fetch the Fedora 38 key is not implemented for
this image at the time. Fedora maintainers indicate it will
make Fedora 39 release. See #6.